+++ packages/jobs/src/define-job/index.ts
@@ -76,6 +76,7 @@ export class Job<
     }
 
     super({
+      runtime: 'job',
       plugins: [],
       handler: invokeHandler,
     })